+++ examples/utest/testcases/tfs/Kconfig
@@ -1,0 +1,6 @@
+menu "Tfs Testcase"
+
+config UTEST_TMPFS_CP
+    bool "tmpfs cp test"
+    default y
+endmenu
