--- examples/utest/testcases/utest copy/Kconfig
@@ -1,7 +1,0 @@
-menu "Utest Self Testcase"
-
-config UTEST_SELF_PASS_TC
-    bool "Pass test"
-    default y
-
-endmenu
